Add support for multiple server queues (failover)
We replace Adhearsion with our own "micro-framework" based on ruby_ami,
which is the underlying Asterisk driver component in Punchblock, the
library that Adhearsion is built upon.

This allows us to connect to multiple servers for failover purposes.
Adhearsion assumes only one server exists, so we can't use it.  Besides,
Adhearsion's main focus is on dial plans, which Queue-tip doesn't use.
Ideally, we'd use Punchblock to support multiple server types, but it's
a big and very complicated dependency, it has an annoying bug and we
don't really need to support anything besides Asterisk.